Add mutex to MutableCFOptions to address the race issue

+++ include/rocksdb/advanced_options.h
@@ -202,10 +202,10 @@ enum class Temperature : uint8_t {
   kCold = 0x0C,
 };
 
-// The control option of how the cache teirs will be used. Currently rocksdb
-// support block cahe (volatile tier), secondary cache (non_vaolatile tier).
+// The control option of how the cache tiers will be used. Currently rocksdb
+// support block cahe (volatile tier), secondary cache (non-volatile tier).
 // In the future, we may add more caching layers.
-enum class CacheTier : unsigned int {
+enum class CacheTier : uint8_t {
   kVolatileTier = 0,
   kNonVolatileTier = 0x01,
 };
